chatbot_server/requirements_chatbot.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.72.0
packaging==24.2
pydantic==2.11.3
pydantic_core==2.33.1
PyPDF2==3.0.1
python-dotenv==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
Werkzeug==3.1.3
zipp==3.21.0
